requirements.txt
flask>=3.0
psycopg2-binary
feedparser
requests
beautifulsoup4
python-dotenv
pyyaml
